// src/modules/exam-configurations/schemas/exam-configuration.schema.ts
import { ExamRule } from '@/modules/exam-rules/entities/exam-rule.entity';
import { Group } from '@/modules/groups/schemas/group.schema';
import { UserRole } from '@/modules/user-roles/schemas/user-role.schemas';
import { Configuration } from '@/shared/schemas/configuration.schema';
import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';

@Schema()
export class ExamConfiguration extends Configuration {
  //boolean to indicate if the exam is timer enabled
  @Prop({ type: Boolean, required: true, default: false })
  isTimerEnabled: boolean;

  //boolean to indicate if the exam is highlighted, or the exam type is highlighted
  //it will have priority in the array order
  @Prop({ type: Boolean, required: true, default: false })
  highlighted: boolean;

  //boolean to indicate if the exam shows answers
  @Prop({ type: Boolean, required: true, default: true })
  showAnswers: boolean;

  //boolean to indicate if the exam is timed
  @Prop({ type: Boolean, required: true, default: false })
  timed: boolean;

  //number to indicate the duration of the exam
  @Prop({ type: Number, required: false })
  duration: number;

  //date to indicate the start time of the exam
  @Prop({ type: Date, required: false })
  startTime: Date;

  //date to indicate the end time of the exam
  @Prop({ type: Date, required: false })
  endTime: Date;

  //date to indicate when the exam will be displayed to the users.
  //Ex: if startTime is 15/07/2025, and displayDate is 14/07/2025, the user will see the exam on 14/07/2025,
  //but will not be able to take it until 15/07/2025
  @Prop({ type: Date, required: false })
  displayDate: Date;

  //number to indicate the number of questions in the exam
  @Prop({ type: Number, required: true, default: 10 })
  numberOfQuestions: number;

  //number to indicate the maximum score of the exam
  @Prop({ type: Number, required: true, default: 10 })
  maxScore: number;

  //number to indicate the maximum attempts of the exam that can be taken
  //if maxAttempts is null, the exam can be taken infinite times
  @Prop({ type: Number, required: false })
  maxAttempts: number;

  //number to indicate the minimum score to pass the exam
  @Prop({ type: Number, required: true, default: 5 })
  passingScore: number;

  //array of exam rules
  @Prop({ type: [{ type: String, required: false, ref: 'ExamRule' }] })
  rules: ExamRule[];

  //allowed roles to take the exam
  @Prop({ type: [{ type: String, required: false, ref: 'UserRole' }] })
  allowedRoles: UserRole[];

  //number to indicate the maximum topics of the exam. If the maxTopis is null, the exam has no topics limit
  @Prop({ type: Number, default: null })
  maxTopics: number;

  //attribute to indicate if the exam ca be autogenerated by an user that is not an admin
  @Prop({ type: Boolean, default: false })
  canAutogenerate: boolean;

  //attribute to indicate the groups that the users must belong to take the exam
  @Prop({ type: [{ type: String, required: false, ref: 'Group' }] })
  groups: Group[];
}
export const ExamConfigurationSchema =
  SchemaFactory.createForClass(ExamConfiguration);